+++ expenseTrackerBackend/services/savings.service.js
@@ -1,10 +1,11 @@
 const { sql, config } = require('../config/db.config');
 
-exports.getAllSavings=async()=>{
+exports.getAllSavings=async(userId)=>{
     try{
         await sql.connect(config);
         const result = await sql.query(`
             SELECT * FROM Savings
+            WHERE userId = ${userId}
             ORDER BY createdAt DESC
         `);
         return result.recordset;
@@ -17,13 +18,14 @@ exports.getAllSavings=async()=>{
 exports.addSavings=async(savingData)=>{
     try{
         await sql.connect(config);
-        const { title, amount, savingsDate } = savingData;
+        const { title, amount, savingsDate,userId } = savingData;
         await sql.query(`
-            INSERT INTO Savings (title, amount, savingsDate)
+            INSERT INTO Savings (title, amount, savingsDate, userId)
             VALUES (
                 '${title}',
                 ${amount},
-                '${savingsDate}'
+                '${savingsDate}',
+                ${userId}
             )
         `);
     }catch(err){
@@ -32,7 +34,7 @@ exports.addSavings=async(savingData)=>{
     }
 }
 
-exports.updateSavings=async(id,savingData)=>{
+exports.updateSavings=async(id,userId,savingData)=>{
     try{
         await sql.connect(config);
         const {title,amount,savingsDate}=savingData;
@@ -42,6 +44,7 @@ exports.updateSavings=async(id,savingData)=>{
             amount=${amount},
             savingsDate='${savingsDate}'
             WHERE id ='${id}'
+            AND userId = ${userId}
             `)
     }catch(err){
         console.error('Error updating saving:', err);
@@ -49,11 +52,11 @@ exports.updateSavings=async(id,savingData)=>{
     }
 }
 
-exports.deleteSavings=async(id)=>{
+exports.deleteSavings=async(id,userId)=>{
     try{
         await sql.connect(config);
         await sql.query(`
-            DELETE FROM Savings WHERE id= ${id}
+            DELETE FROM Savings WHERE id= ${id} AND userId = ${userId}
             `)
     }catch(err){
         console.error('Error deleting saving:', err);
